Changed all fields relating to money from float to decimal as it is better for financial units

+++ backend/Controllers/GroupsController.cs
@@ -44,7 +44,7 @@ namespace backend.Controllers
         }
         // GET: api/GroupDebt/5
         [HttpGet("GroupDebt/{id:int}")]
-        public async Task<ActionResult<float>> GetGroupDebt(int id)
+        public async Task<ActionResult<decimal>> GetGroupDebt(int id)
         {
             if (!ModelState.IsValid)
             {
